loader: Viewport width is set to document width
When data-framework-viewport-width is set to screen-width,
the viewport was set to window.outerWidth. However, with web
simulator, webapps are put in the iframe, so the viewport size
needs to be set to document width. This patch sets the viewport
width to document.documentElement.clientWidth, as
document.documentElement actually means <html> element.

Change-Id: I4c1d4002e7d276ad274681afab484942e8069abf